\name{CommonThreads-package}
\alias{CommonThreads-package}
\alias{CommonThreads}
\docType{package}
\title{Which variables best explain membership?}
\description{
Given a (rectangular) data set and some subset of (the rows of) the data, identifies which columns (variables) best predict membership in that subset. One way to interpret this is that these variables provide the "common thread" among the observations in the subset.
}
\details{
\tabular{ll}{
Package: \tab CommonThreads\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-07\cr
License: \tab MIT + LICENSE\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Stephen R. Haptonstahl <srh@haptonstahl.org>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
